% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_intervals.R
\name{anno-interval}
\alias{anno-interval}
\alias{anno_errorbar}
\alias{anno_linerange}
\title{Intervals annotation: lines, crossbars & errorbars}
\usage{
anno_errorbar(
  x,
  which = NULL,
  width = 0.9,
  pch = 19,
  size = 1,
  point_gp = gpar(),
  line_gp = gpar(),
  band = TRUE,
  band_col = c("white", "#eff3f2")
)

anno_linerange(
  x,
  which = NULL,
  pch = 19,
  size = 1,
  point_gp = gpar(),
  line_gp = gpar(),
  band = TRUE,
  band_col = c("white", "#eff3f2")
)
}
\arguments{
\item{x}{A matrix with at leat 3 columns, regarded as the error bar point
estimate and confidence interval (lower and higher).}

\item{which}{Whether it is drawn as a column annotation or a row annotation?}

\item{width}{A \link[grid:unit]{unit} object specifying the width of the tip
head.}

\item{pch}{numeric or character vector indicating what sort of
    plotting symbol to use.  See \code{\link{points}} for the
    interpretation of these values, and note \code{fill} below.}

\item{size}{unit object specifying the size of the plotting symbols.}

\item{point_gp}{A \link[grid:gpar]{gpar} objects define the point attributes.}

\item{line_gp}{A \link[grid:gpar]{gpar} objects define the line attributes.}

\item{band}{A bool, if \code{TRUE}, will add color band interval}

\item{band_col}{Color define the band interval. Should be of length 2L.}
}
\description{
Intervals annotation: lines, crossbars & errorbars
}
