% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_factory.R
\name{layer-fun}
\alias{layer-fun}
\alias{layer_band}
\alias{layer_text}
\title{Function factories for \code{layer_fun} in ComplexHeatmap}
\usage{
layer_band(band, band_col = c("white", "#eff3f2"))

layer_text(matrix, format = NULL, gp = gpar())
}
\arguments{
\item{band}{A string, "row" or "column", add row or column interval color
band.}

\item{band_col}{Color define the band interval. Should be of length 2L.}

\item{matrix}{Text matrix used to draw.}

\item{format}{Format strings, see \link{sprintf}.}

\item{gp}{A \link[grid:gpar]{gpar} objects define the text attributes.}
}
\value{
A layer_fun function for the usage of
\link[ComplexHeatmap:Heatmap]{Heatmap}.
}
\description{
Function factories for \code{layer_fun} in ComplexHeatmap
}
\details{
\itemize{
\item \code{layer_band}: draw color interval band.
\item \code{layer_text}: draw text.
}
}
